Move extension methods around
